Separate inner class declarations
