+++ ngrinder-groovy/src/main/java/org/ngrinder/http/ThreadContextCookieJar.java
@@ -1,0 +1,41 @@
+/*
+ * Copyright (c) 2012-present NAVER Corp.
+ *
+ * This file is part of The nGrinder software distribution. Refer to
+ * the file LICENSE which is part of The nGrinder distribution for
+ * licensing details. The nGrinder distribution is available on the
+ * Internet at https://naver.github.io/ngrinder
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.ngrinder.http;
+
+import okhttp3.Cookie;
+import okhttp3.CookieJar;
+import okhttp3.HttpUrl;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
+
+public class ThreadContextCookieJar implements CookieJar {
+	@Override
+	public void saveFromResponse(@NotNull HttpUrl httpUrl, @NotNull List<Cookie> list) {
+		CookieManager.addCookies(list);
+	}
+
+	@NotNull
+	@Override
+	public List<Cookie> loadForRequest(@NotNull HttpUrl httpUrl) {
+		return CookieManager.getCookies();
+	}
+}
